Added some changes requested 1. Changed the comments of the flattened dictionary 2. Corrected some typo for fits_checker 3. Changed the name of the `to_hdu` to `to_hdulist` 4. Changed the name of the ImageHDU list 5. Used `isinstance` to conform to numpy mask 6. Changed the LABEL for extra coords 7. Shifted the `fits_ndcube_writer`
